Quick advances

+++ src/model/ContaCorrente.java
@@ -1,7 +1,7 @@
 package src.model;
 
 public class ContaCorrente extends Conta{
-    public ContaCorrente(int agencia, long numero, double saldo){
+    public ContaCorrente(int agencia, int numero, double saldo){
         super(agencia, numero, saldo);
     }
 }
