+++ src/model/ContaPoupanca.java
@@ -1,7 +1,7 @@
 package src.model;
 
 public class ContaPoupanca extends Conta{
-    public ContaPoupanca(int agencia, long numero, double saldo){
+    public ContaPoupanca(int agencia, int numero, double saldo){
         super(agencia, numero, saldo);
     }
 }
